Move functions from utils
